requirements.txt
python-telegram-bot>=20.7
yt-dlp>=2024.1.0
pytumblr>=0.1.0
python-dotenv>=1.0.0
requests-oauthlib>=1.3.1
imageio-ffmpeg>=0.5.1
